fix(windows): rename header - review comments
