+++ windows/src/engine/keyman32/kmhook_getmessage.cpp
@@ -65,7 +65,7 @@
 #include "pch.h"
 
 void ProcessWMKeymanControlInternal(HWND hwnd, WPARAM wParam, LPARAM lParam);
-void ProcessWMKeymanControl(HWND hwnd, WPARAM wParam, LPARAM lParam);
+void ProcessWMKeymanControl(WPARAM wParam, LPARAM lParam);
 void ProcessWMKeyman(HWND hwnd, WPARAM wParam, LPARAM lParam);
 void GetCapsAndNumlockState();
 
@@ -237,7 +237,7 @@ LRESULT _kmnGetMessageProc(int nCode, WPARAM wParam, LPARAM lParam)
   if (mp->message == wm_keyman_control)
   {
       SendDebugMessageFormat(0, sdmInternat, 0, "GetMessage: wm_keyman_control hwnd=%x %x %x", mp->hwnd, mp->wParam, mp->lParam);
-      ProcessWMKeymanControl(mp->hwnd, mp->wParam, mp->lParam);
+      ProcessWMKeymanControl(mp->wParam, mp->lParam);
       return CallNextHookEx(Globals::get_hhookGetMessage(), nCode, wParam, lParam);
   }
 
@@ -421,8 +421,7 @@ void ProcessWMKeymanControlInternal(HWND hwnd, WPARAM wParam, LPARAM lParam)
 }
 
 void
-ProcessWMKeymanControl(HWND hwnd, WPARAM wParam, LPARAM lParam) {
-  UNREFERENCED_PARAMETER(hwnd);
+ProcessWMKeymanControl(WPARAM wParam, LPARAM lParam) {
 
   switch (wParam) {
   case KMC_PROFILECHANGED:
