+++ windows/src/engine/keyman64/keyman64.vcxproj
@@ -175,7 +175,7 @@
     </Bscmake>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\global\vc\keymankeyboard_id.cpp" />
+    <ClCompile Include="..\..\global\vc\kmtip_guids.cpp" />
     <ClCompile Include="..\..\global\vc\xstring.cpp" />
     <ClCompile Include="..\keyman32\addins.cpp" />
     <ClCompile Include="..\keyman32\appint\aiTIP.cpp" />
@@ -283,7 +283,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\global\inc\keymanengine.h" />
-    <ClInclude Include="..\..\global\inc\keymankeyboard_id.h" />
+    <ClInclude Include="..\..\global\inc\kmtip_guids.h" />
     <ClInclude Include="..\keyman32\addins.h" />
     <ClInclude Include="..\keyman32\appint\aiTIP.h" />
     <ClInclude Include="..\keyman32\appint\aiWin2000Unicode.h" />
