+++ windows/src/engine/keyman64/keyman64.vcxproj.filters
@@ -38,7 +38,7 @@
     <ClCompile Include="..\keyman32\VKScanCodes.cpp" />
     <ClCompile Include="..\keyman32\DebugEventTrace.cpp" />
     <ClCompile Include="..\keyman32\kmprocessactions.cpp" />
-    <ClCompile Include="..\..\global\vc\keymankeyboard_id.cpp" />
+    <ClCompile Include="..\..\global\vc\kmtip_guids.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\keyman32\addins.h" />
@@ -87,7 +87,7 @@
     <ClInclude Include="..\keyman32\vkscancodes.h" />
     <ClInclude Include="..\..\global\inc\keymanengine.h" />
     <ClInclude Include="..\keyman32\kmprocessactions.h" />
-    <ClInclude Include="..\..\global\inc\keymankeyboard_id.h" />
+    <ClInclude Include="..\..\global\inc\kmtip_guids.h" />
   </ItemGroup>
   <ItemGroup>
     <CustomBuild Include="..\keyman32\keyman-debug-etw.man" />
