+++ windows/src/engine/kmtip/kmtip.vcxproj
@@ -275,7 +275,7 @@
     </Bscmake>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="..\..\global\vc\keymankeyboard_id.cpp" />
+    <ClCompile Include="..\..\global\vc\kmtip_guids.cpp" />
     <ClCompile Include="debug.cpp" />
     <ClCompile Include="keyman32interface.cpp" />
     <ClCompile Include="keys.cpp">
@@ -417,7 +417,7 @@
   <ItemGroup>
     <ClInclude Include="..\..\global\inc\certificate_check.h" />
     <ClInclude Include="..\..\global\inc\Compiler.h" />
-    <ClInclude Include="..\..\global\inc\keymankeyboard_id.h" />
+    <ClInclude Include="..\..\global\inc\kmtip_guids.h" />
     <ClInclude Include="editsess.h" />
     <ClInclude Include="globals.h" />
     <ClInclude Include="..\..\global\inc\KEYNAMES.H" />
