+++ windows/src/engine/kmtip/kmtip.vcxproj.filters
@@ -69,7 +69,7 @@
     <ClCompile Include="keyman32interface.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\..\global\vc\keymankeyboard_id.cpp">
+    <ClCompile Include="..\..\global\vc\kmtip_guids.cpp">
       <Filter>Source Files\Framework</Filter>
     </ClCompile>
   </ItemGroup>
@@ -111,7 +111,7 @@
     <ClInclude Include="pch.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="..\..\global\inc\keymankeyboard_id.h">
+    <ClInclude Include="..\..\global\inc\kmtip_guids.h">
       <Filter>Header Files</Filter>
     </ClInclude>
   </ItemGroup>
