+++ windows/src/global/inc/kmtip_guids.h
@@ -1,0 +1,13 @@
+/*
+ * Keyman is copyright (C) SIL International. MIT License.
+ *
+ * Name:             kmtip_guids.h
+ * Description:      Header file to include the Keyman keyboard GUIDs
+ */
+#ifndef _kmtip_guids_id_h
+#define _kmtip_guids_id_h
+
+/* FE0420F1-38D1-4B4C-96BF-E7E20A74CFB7 kmtip (Keyman Text Input Processor) text service CLSID */
+extern const CLSID c_clsidKMTipTextService;
+
+#endif
