+++ windows/src/global/vc/kmtip_guids.cpp
